requirements.txt
celery==5.2.7
click==8.1.3
cryptography==39.0.1
Django==4.1.9
djangorestframework==3.14.0
django-cors-headers==3.13.0
numpy==1.22.0
psutil==5.9.2
pydantic==1.10.2
pymongo==4.2.0
PyYAML==6.0
redis==4.4.4
requests==2.31.0
scikit-learn==1.0.2
sentry-sdk==1.19.1
tqdm==4.64.1
whitenoise==6.2.0
